Move proxy configuration form out of pluginManager screens as it is not related
