+++ core/src/main/java/hudson/PluginManager.java
@@ -45,7 +45,6 @@ import hudson.model.AbstractItem;
 import hudson.model.AbstractModelObject;
 import hudson.model.AdministrativeMonitor;
 import hudson.model.Api;
-import hudson.model.Descriptor;
 import hudson.model.DownloadService;
 import hudson.model.Failure;
 import hudson.model.ItemGroupMixIn;
@@ -166,7 +165,6 @@ import org.kohsuke.stapler.StaplerResponse;
 import org.kohsuke.stapler.export.Exported;
 import org.kohsuke.stapler.export.ExportedBean;
 import org.kohsuke.stapler.interceptor.RequirePOST;
-import org.kohsuke.stapler.verb.POST;
 import org.springframework.security.core.Authentication;
 import org.xml.sax.Attributes;
 import org.xml.sax.InputSource;
@@ -1797,22 +1795,6 @@ public abstract class PluginManager extends AbstractModelObject implements OnMas
         return new HttpRedirect("advanced");
     }
 
-    @POST
-    public HttpResponse doProxyConfigure(StaplerRequest req) throws IOException, ServletException {
-        Jenkins jenkins = Jenkins.get();
-        jenkins.checkPermission(Jenkins.ADMINISTER);
-
-        ProxyConfiguration pc = req.bindJSON(ProxyConfiguration.class, req.getSubmittedForm());
-        if (pc.name == null) {
-            jenkins.proxy = null;
-            ProxyConfiguration.getXmlFile().delete();
-        } else {
-            jenkins.proxy = pc;
-            jenkins.proxy.save();
-        }
-        return new HttpRedirect("advanced");
-    }
-
     interface PluginCopier {
         void copy(File target) throws Exception;
 
@@ -2128,10 +2110,6 @@ public abstract class PluginManager extends AbstractModelObject implements OnMas
             LOGGER.log(WARNING, "Failed to identify the short name from " + t, e);
         }
         return FilenameUtils.getBaseName(t.getName());    // fall back to the base name of what's uploaded
-    }
-
-    public Descriptor<ProxyConfiguration> getProxyDescriptor() {
-        return Jenkins.get().getDescriptor(ProxyConfiguration.class);
     }
 
     /**
