+++ core/src/main/resources/hudson/Messages.properties
@@ -127,3 +127,5 @@ PluginWrapper.NoSuchPlugin=No such plugin found with the name ''{0}''
 PluginWrapper.Error.Disabling=There was an error disabling the ''{0}'' plugin. Error: ''{1}''
 TcpSlaveAgentListener.PingAgentProtocol.displayName=Ping protocol
 
+ProxyConfigurationManager.DisplayName=Proxy Configuration
+ProxyConfigurationManager.Description=Configure the http proxy used by Jenkins
