+++ core/src/main/resources/hudson/PluginManager/advanced.jelly
@@ -38,21 +38,6 @@ THE SOFTWARE.
           <l:app-bar title="${%Advanced settings}"/>
 
           <div class="jenkins-form">
-            <section class="jenkins-section jenkins-!-margin-bottom-5">
-                <h2 class="jenkins-section__title">${%HTTP Proxy Configuration}</h2>
-                <f:form method="post" action="proxyConfigure" name="proxyConfigure">
-                    <j:scope>
-                        <j:set var="instance" value="${app.proxy}"/>
-                        <j:set var="descriptor" value="${it.proxyDescriptor}"/>
-                        <st:include from="${descriptor}" page="${descriptor.configPage}"/>
-                    </j:scope>
-                    <l:isAdmin>
-                        <f:block>
-                            <f:submit/>
-                        </f:block>
-                    </l:isAdmin>
-                </f:form>
-            </section>
 
             <l:isAdmin>
                 <section class="jenkins-section jenkins-!-margin-bottom-5">
