+++ core/src/main/resources/hudson/ProxyConfigurationManager/index.jelly
@@ -1,0 +1,50 @@
+<!--
+The MIT License
+
+Copyright (c) 2023, CloudBees Inc, and other contributors
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+-->
+
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
+         xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+    <l:layout title="${%Proxy Configuration}" permission="${app.SYSTEM_READ}">
+
+        <l:main-panel>
+          <div class="jenkins-form">
+            <section class="jenkins-section jenkins-!-margin-bottom-5">
+                <h2 class="jenkins-section__title">${%HTTP Proxy Configuration}</h2>
+                <f:form method="post" action="proxyConfigure" name="proxyConfigure">
+                    <j:scope>
+                        <j:set var="instance" value="${app.proxy}"/>
+                        <j:set var="descriptor" value="${it.proxyDescriptor}"/>
+                        <st:include from="${descriptor}" page="${descriptor.configPage}"/>
+                    </j:scope>
+                    <l:isAdmin>
+                        <f:block>
+                            <f:submit/>
+                        </f:block>
+                    </l:isAdmin>
+                </f:form>
+            </section>
+          </div>
+        </l:main-panel>
+    </l:layout>
+</j:jelly>
